chore: add code quality tooling and CI pipeline - Add clang-tidy, scan-build, debug sanitizers, valgrind targets - Add pre-commit hook (format-check + debug build) - Add GitHub Actions CI (build, debug, tidy, analyze, format-check) - Replace make lint with format-check + build + tidy - Fix clang-tidy findings (atoi→strtol, float loop counter, narrowing casts, macro parens)

+++ src/spectrum.c
@@ -336,7 +336,7 @@ spectrum_init(spectrum_state_t *s, Wave *wave, Font font)
     s->hop_size = FFT_HOP_SIZE;
     s->seconds_per_window = (f64)s->hop_size / (f64)wave->sampleRate;
     s->font = font;
-    s->sample_rate = wave->sampleRate;
+    s->sample_rate = (i32)wave->sampleRate;
 
     for (i32 i = 0; i < FFT_WINDOW_SIZE; i++)
     {
